+++ src/app/components/search/search-results/search-results.component.html
@@ -3,78 +3,9 @@
         <adf-breadcrumb root="APP.BROWSE.SEARCH.TITLE">
         </adf-breadcrumb>
         <adf-toolbar class="inline">
-            <!-- <ng-container *ifExperimental="'extensions'"> -->
             <ng-container *ngFor="let entry of actions; trackBy: trackByActionId">
                 <aca-toolbar-action [entry]="entry"></aca-toolbar-action>
             </ng-container>
-            <!-- </ng-container> -->
-
-            <!--
-            <ng-container *ngIf="!selection.isEmpty">
-                <button
-                    color="primary"
-                    mat-icon-button
-                    *ngIf="selection.file"
-                    title="{{ 'APP.ACTIONS.VIEW' | translate }}"
-                    (click)="showPreview(selection.file)">
-                    <mat-icon>open_in_browser</mat-icon>
-                </button>
-
-                <button
-                    color="primary"
-                    mat-icon-button
-                    title="{{ 'APP.ACTIONS.DOWNLOAD' | translate }}"
-                    (click)="downloadSelection()">
-                    <mat-icon>get_app</mat-icon>
-                </button>
-
-                <app-toggle-info-drawer></app-toggle-info-drawer>
-
-                <button
-                    color="primary"
-                    mat-icon-button
-                    title="{{ 'APP.ACTIONS.MORE' | translate }}"
-                    [matMenuTriggerFor]="actionsMenu">
-                    <mat-icon>more_vert</mat-icon>
-                </button>
-
-                <mat-menu #actionsMenu="matMenu" [overlapTrigger]="false">
-                    <button
-                        mat-menu-item
-                        #favorites="adfFavorite"
-                        [adf-node-favorite]="selection.nodes">
-                        <mat-icon color="primary" *ngIf="favorites.hasFavorites()">star</mat-icon>
-                        <mat-icon *ngIf="!favorites.hasFavorites()">star_border</mat-icon>
-                        <span>{{ 'APP.ACTIONS.FAVORITE' | translate }}</span>
-                    </button>
-
-                    <button
-                        mat-menu-item
-                        [acaCopyNode]="selection.nodes">
-                        <mat-icon>content_copy</mat-icon>
-                        <span>{{ 'APP.ACTIONS.COPY' | translate }}</span>
-                    </button>
-
-                    <button
-                        mat-menu-item
-                        *ngIf="selection.file"
-                        [acaNodeVersions]="selection.file">
-                        <mat-icon>history</mat-icon>
-                        <span>{{ 'APP.ACTIONS.VERSIONS' | translate }}</span>
-                    </button>
-
-                    <ng-container *ifExperimental="'permissions'">
-                        <button
-                            mat-menu-item
-                            *ngIf="canUpdateNode"
-                            [acaNodePermissions]="selection.first">
-                            <mat-icon>settings_input_component</mat-icon>
-                            <span>{{ 'APP.ACTIONS.PERMISSIONS' | translate }}</span>
-                        </button>
-                    </ng-container>
-                </mat-menu>
-            </ng-container>
-            -->
         </adf-toolbar>
     </div>
